+++ examples/20-transform-kernels.cpp
@@ -54,7 +54,7 @@ int main() {
   auto matrix = Grid(lMax, n, FFTWpp::Measure, Chunking::Automatic(),
                      WignerValues::Stored(), TransformKernel::Matrix());
 
-  const auto fieldSize = static_cast<Int>(loop.FieldSize());
+  const auto fieldSize = loop.FieldSize();
   const auto coefficientSize = static_cast<Int>(loop.CoefficientSize(lMax, n));
 
   auto fields = std::vector<Complex>(count * fieldSize);
